Fix build with double precision mode

+++ src/liblightmetrica/asset/texture/texture_bitmap.cpp
@@ -143,7 +143,7 @@ public:
             // --------------------------------------------------------------------------------
 
             // Scale
-            const auto scale = prop->ChildAs<Float>("scale", 1_f);
+            const auto scale = (float)(prop->ChildAs<Float>("scale", 1_f));
             for (auto& v : data_)
             {
                 v *= scale;
